Consolidated the configuration library into the application library
Closes #48
