+++ src/Application/composer.json
@@ -25,7 +25,8 @@
         }
     },
     "require": {
-        "php": "^7.4"
+        "php": "^7.4",
+        "symfony/dotenv": "~5.1"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.0",
