+++ src/Application/src/Configuration/GlobalConfiguration.php
@@ -10,7 +10,7 @@
 
 declare(strict_types=1);
 
-namespace Aphiria\Configuration;
+namespace Aphiria\Application\Configuration;
 
 use RuntimeException;
