+++ src/Application/src/Configuration/HashTableConfiguration.php
@@ -10,7 +10,7 @@
 
 declare(strict_types=1);
 
-namespace Aphiria\Configuration;
+namespace Aphiria\Application\Configuration;
 
 /**
  * Defines a wrapper around a hash table configuration
