+++ src/Application/tests/Configuration/HashTableConfigurationTest.php
@@ -10,10 +10,10 @@
 
 declare(strict_types=1);
 
-namespace Aphiria\Configuration\Tests;
+namespace Aphiria\Application\Tests\Configuration;
 
-use Aphiria\Configuration\HashTableConfiguration;
-use Aphiria\Configuration\MissingConfigurationValueException;
+use Aphiria\Application\Configuration\HashTableConfiguration;
+use Aphiria\Application\Configuration\MissingConfigurationValueException;
 use PHPUnit\Framework\TestCase;
 
 class HashTableConfigurationTest extends TestCase
