+++ src/Application/tests/Configuration/PhpConfigurationFileReaderTest.php
@@ -10,10 +10,10 @@
 
 declare(strict_types=1);
 
-namespace Aphiria\Configuration\Tests;
+namespace Aphiria\Application\Tests\Configuration;
 
-use Aphiria\Configuration\InvalidConfigurationFileException;
-use Aphiria\Configuration\PhpConfigurationFileReader;
+use Aphiria\Application\Configuration\InvalidConfigurationFileException;
+use Aphiria\Application\Configuration\PhpConfigurationFileReader;
 use PHPUnit\Framework\TestCase;
 
 class PhpConfigurationFileReaderTest extends TestCase
