+++ src/Application/tests/Configuration/files/configuration-delimiter.json
