+++ src/Application/tests/Configuration/files/invalid-configuration.php
