--- src/Configuration/.gitattributes
@@ -1,10 +1,0 @@
-# Auto detect text files and perform LF normalization
-* text=lf
-
-# Ignore select directories on export
-.gitattributes export-ignore
-.gitignore export-ignore
-CONTRIBUTING.md export-ignore
-phpunit.xml export-ignore
-README.md export-ignore
-/tests export-ignore
