--- src/Configuration/.travis.yml
@@ -1,27 +1,0 @@
-language: php
-
-php:
-  - 7.4
-
-before_script:
-  - if php -v | grep -q 'Xdebug'; then phpenv config-rm xdebug.ini; fi
-  - pecl install pcov
-  - composer self-update
-  - composer install
-  - composer dump-autoload -o
-
-script:
-  - vendor/bin/phpunit --coverage-clover ./.coverage/clover.xml --verbose
-  - composer run-script lint-check
-
-after_success:
-  - vendor/bin/php-coveralls --coverage_clover=./.coverage/clover.xml --json_path=./coveralls-upload.json -v
-
-notifications:
-  email:
-    - dave@aphiria.com
-
-sudo: false
-
-matrix:
-  fast_finish: true
