--- src/Configuration/CONTRIBUTING.md
@@ -1,3 +1,0 @@
-# Contributing to Aphiria
-
-This repository is a read-only subtree split, and should not be contributed to directly.  Please review the [contribution guide](https://www.aphiria.com/docs/master/contributing.html) for details on how to contribute.
