--- src/Configuration/README.md
@@ -1,25 +1,0 @@
-<p align="center"><a href="https://www.aphiria.com" target="_blank" title="Aphiria"><img src="https://www.aphiria.com/images/aphiria-logo.svg"></a></p>
-
-<p align="center">
-<a href="https://travis-ci.com/aphiria/configuration"><img src="https://travis-ci.com/aphiria/configuration.svg?branch=master"></a>
-<a href='https://coveralls.io/github/aphiria/configuration?branch=master'><img src='https://coveralls.io/repos/github/aphiria/configuration/badge.svg?branch=master' alt='Coverage Status' /></a>
-<a href="https://packagist.org/packages/aphiria/configuration"><img src="https://poser.pugx.org/aphiria/configuration/v/stable.svg"></a>
-<a href="https://packagist.org/packages/aphiria/configuration"><img src="https://poser.pugx.org/aphiria/configuration/v/unstable.svg"></a>
-<a href="https://packagist.org/packages/aphiria/configuration"><img src="https://poser.pugx.org/aphiria/configuration/license.svg"></a>
-</p>
-
-# Configuration
-
-> **Note:** This library is not stable yet.
-
-## Installation
-
-This library can be installed via [Composer](https://getcomposer.org/download/):
-
-```php
-composer require aphiria/configuration:^1.0@dev
-```
-
-## Documentation
-
-Full documentation is available <a href="https://www.aphiria.com/docs/master/configuration.html" target="_blank">here</a>.
