--- src/Configuration/composer.json
@@ -1,45 +1,0 @@
-{
-    "name": "aphiria/configuration",
-    "description": "The Aphiria application configuration library",
-    "keywords": [
-        "PHP",
-        "application",
-        "configuration"
-    ],
-    "authors": [
-        {
-            "name": "David Young",
-            "email": "dave@aphiria.com",
-            "role": "Architect"
-        }
-    ],
-    "license": "MIT",
-    "autoload": {
-        "psr-4": {
-            "Aphiria\\Configuration\\": "src"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "Aphiria\\Configuration\\Tests\\": "tests"
-        }
-    },
-    "require": {
-        "php": "^7.4"
-    },
-    "require-dev": {
-        "friendsofphp/php-cs-fixer": "^3.0",
-        "php-coveralls/php-coveralls": "^2.2",
-        "phpunit/phpunit": "~9.1"
-    },
-    "scripts": {
-        "lint-check": "php-cs-fixer fix --config=.php_cs.dist -v --diff --dry-run",
-        "lint-fix": "php-cs-fixer fix --config=.php_cs.dist -v --diff"
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "1.0.x-dev"
-        }
-    },
-    "minimum-stability": "dev"
-}
