--- src/Configuration/phpunit.xml.dist
@@ -1,15 +1,0 @@
-<phpunit
-    bootstrap="vendor/autoload.php"
-    colors="true">
-    <testsuite name="Configuration">
-        <directory>tests</directory>
-    </testsuite>
-    <filter>
-        <whitelist>
-            <directory suffix=".php">src</directory>
-        </whitelist>
-    </filter>
-    <logging>
-        <log type="coverage-html" target="./.coverage"/>
-    </logging>
-</phpunit>
