+++ src/Framework/src/Exceptions/Bootstrappers/GlobalExceptionHandlerBootstrapper.php
@@ -15,9 +15,9 @@ namespace Aphiria\Framework\Exceptions\Bootstrappers;
 use Aphiria\Api\Errors\ProblemDetailsResponseMutator;
 use Aphiria\Api\Validation\InvalidRequestBodyException;
 use Aphiria\Api\Validation\ValidationProblemDetails;
+use Aphiria\Application\Configuration\GlobalConfiguration;
+use Aphiria\Application\Configuration\MissingConfigurationValueException;
 use Aphiria\Application\IBootstrapper;
-use Aphiria\Configuration\GlobalConfiguration;
-use Aphiria\Configuration\MissingConfigurationValueException;
 use Aphiria\DependencyInjection\IContainer;
 use Aphiria\Exceptions\GlobalExceptionHandler;
 use Aphiria\Exceptions\IExceptionRenderer;
