+++ src/Framework/tests/Net/Binders/ContentNegotiationBinderTest.php
@@ -12,8 +12,8 @@ declare(strict_types=1);
 
 namespace Aphiria\Framework\Tests\Net\Binders;
 
-use Aphiria\Configuration\GlobalConfiguration;
-use Aphiria\Configuration\HashTableConfiguration;
+use Aphiria\Application\Configuration\GlobalConfiguration;
+use Aphiria\Application\Configuration\HashTableConfiguration;
 use Aphiria\DependencyInjection\IContainer;
 use Aphiria\Framework\Net\Binders\ContentNegotiationBinder;
 use Aphiria\Net\Http\ContentNegotiation\AcceptCharsetEncodingMatcher;
