+++ src/Framework/tests/Routing/Binders/RoutingBinderTest.php
@@ -12,8 +12,8 @@ declare(strict_types=1);
 
 namespace Aphiria\Framework\Tests\Routing\Binders;
 
-use Aphiria\Configuration\GlobalConfiguration;
-use Aphiria\Configuration\HashTableConfiguration;
+use Aphiria\Application\Configuration\GlobalConfiguration;
+use Aphiria\Application\Configuration\HashTableConfiguration;
 use Aphiria\DependencyInjection\IContainer;
 use Aphiria\Framework\Routing\Binders\RoutingBinder;
 use Aphiria\Routing\Annotations\AnnotationRouteRegistrant;
