+++ src/Framework/tests/Serialization/Binders/SymfonySerializerBinderTest.php
@@ -12,8 +12,8 @@ declare(strict_types=1);
 
 namespace Aphiria\Framework\Tests\Serialization\Binders;
 
-use Aphiria\Configuration\GlobalConfiguration;
-use Aphiria\Configuration\HashTableConfiguration;
+use Aphiria\Application\Configuration\GlobalConfiguration;
+use Aphiria\Application\Configuration\HashTableConfiguration;
 use Aphiria\DependencyInjection\IContainer;
 use Aphiria\Framework\Serialization\Binders\SymfonySerializerBinder;
 use PHPUnit\Framework\MockObject\MockObject;
