+++ src/Framework/tests/Validation/Binders/ValidationBinderTest.php
@@ -12,8 +12,8 @@ declare(strict_types=1);
 
 namespace Aphiria\Framework\Tests\Validation\Binders;
 
-use Aphiria\Configuration\GlobalConfiguration;
-use Aphiria\Configuration\HashTableConfiguration;
+use Aphiria\Application\Configuration\GlobalConfiguration;
+use Aphiria\Application\Configuration\HashTableConfiguration;
 use Aphiria\DependencyInjection\IContainer;
 use Aphiria\Framework\Validation\Binders\ValidationBinder;
 use Aphiria\Validation\Constraints\Annotations\AnnotationObjectConstraintsRegistrant;
